move functions to `utils.js` to unittest sanitizeFileName
